backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["G0dles43 <85917@student.pb.edu.pl>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "5.1.4"
django-cors-headers = "4.6.0"
asgiref = "3.8.1"
build = "1.2.2.post1"
cachecontrol = "0.14.1"
certifi = "2024.12.14"
charset-normalizer = "3.4.1"
cleo = "2.1.0"
click = "8.1.8"
colorama = "0.4.6"
contourpy = "1.2.1"
crashtest = "0.4.1"
cycler = "0.12.1"
distlib = "0.3.9"
django-ranged-response = "0.2.0"
django-simple-captcha = "0.6.0"
djangorestframework = "3.15.2"
dulwich = "0.22.7"
fastjsonschema = "2.21.1"
filelock = "3.16.1"
fonttools = "4.51.0"
idna = "3.10"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.1.0"
keyring = "25.6.0"
kiwisolver = "1.4.5"
lxml = "5.3.0"
markdown-it-py = "3.0.0"
matplotlib = "3.8.4"
mdurl = "0.1.2"
more-itertools = "10.5.0"
msgpack = "1.1.0"
numpy = "1.26.4"
packaging = "24.2"
pexpect = "4.9.0"
pkginfo = "1.12.0"
platformdirs = "4.3.6"
poetry = "2.0.0"
poetry-core = "2.0.0"
poetry-plugin-export = "1.8.0"
ptyprocess = "0.7.0"
pygments = "2.19.1"
pyparsing = "3.1.2"
pyproject-hooks = "1.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.11.0"
reportlab = "4.2.0"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
shellingham = "1.5.4"
six = "1.16.0"
sqlparse = "0.5.2"
tomlkit = "0.13.2"
trove-classifiers = "2024.10.21.16"
typing-extensions = "4.11.0"
tzdata = "2024.1"
urllib3 = "2.3.0"
virtualenv = "20.28.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
drf-spectacular = "^0.27.2"
django-extensions = "^3.2.3"
axios = "0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
